// lib/types.ts
export interface Lab {
  id: string;
  name: string;
  address: string;
  city: string;
  state: string;
  website: string;
  phone: string;
  email: string;
  services: string[];
  formats: string[];
  developOnly35mm: string;
  developScan35mm: string;
  developHiRes35mm: string;
  developPrint35mm: string;
  expressService: string;
  turnaround: string;
  specialServices: string;
  mailIn: string;
  notes: string;
}